Skip ListPopupWindow position update if detached
Symptom:
A part of the option is displayed in the wrong location.

Root cause:
When the search item is clicked, the menu item will be
hidden. The PopupWindow will be detached from window and the
mAttachInfo property will be set to null. The anchor coordinates
will return x 0 and y 0.

Solution:
If the anchor view is detached from window, the ListPopupWindow will
not be update.

Bug: 36207919

Change-Id: I2b370c31542fae454592463e12313fdeaef25afe

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>